feat: expanded Director() and created Config()
The Director() will be the main class responsible for manipulating the file system. To create, read,
or delete files you can use the Director(). The Config() class extends the ConfigBase() class. The
ConfigBase() class maintains the config and contains functions to read and update the config.json.
The Config() class can be extended by the user to create/edit/update config values